Rename module from 'parser' to 'dacot_parser'
